refactor: split fallback next-turn, corner controls, quick battle (#3878)
Extract isolate-backed fallback turn-resolution into a GameScreen part,
split GameMapCornerControls icon-button chrome into a library part, and
move QuickBattleScreen result summary into a library part without API changes.

Refs #3878
