+++ app/lib/features/game/screens/combat/quick_battle_screen.dart
@@ -10,6 +10,8 @@ import '../../../../widgets/ct_nine_patch_button.dart';
 import '../../widgets/combat/quick_battle_action_selector.dart';
 import '../../widgets/combat/quick_battle_deployment_view.dart';
 
+part 'quick_battle_screen_result.dart';
+
 /// Quick Battle flow: deployment → rounds → result. SPEC/game/quick-battle.md.
 /// Uses default actions (Volley Fire) when run in headless/AI mode.
 class QuickBattleScreen extends StatefulWidget {
@@ -117,74 +119,6 @@ class _QuickBattleScreenState extends State<QuickBattleScreen> {
               child: Text(l10n.quickBattle_resolveAuto),
             ),
           ],
-        ],
-      ),
-    );
-  }
-}
-
-class _ResultView extends StatelessWidget {
-  const _ResultView({required this.result, required this.onDismiss});
-
-  final QuickBattleResult result;
-  final VoidCallback onDismiss;
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = appL10n(context);
-    final theme = Theme.of(context);
-    final winnerText = switch (result.winner) {
-      QuickBattleWinner.attacker => l10n.quickBattle_attackerWins(
-        l10n.quickBattle_attackerDefaultName,
-      ),
-      QuickBattleWinner.defender => l10n.quickBattle_defenderHolds(
-        l10n.quickBattle_defenderDefaultName,
-      ),
-      QuickBattleWinner.mutualExhaustion => l10n.quickBattle_mutualExhaustion,
-    };
-    return CtDialogShell(
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(
-            l10n.quickBattle_battleResult(winnerText),
-            style: theme.textTheme.titleMedium?.copyWith(
-              color: theme.colorScheme.primary,
-            ),
-          ),
-          const SizedBox(height: 8),
-          if (result.provinceFlips)
-            Text(
-              l10n.quickBattle_provinceCaptured,
-              style: TextStyle(
-                fontWeight: FontWeight.bold,
-                color: theme.colorScheme.error,
-              ),
-            ),
-          const SizedBox(height: 8),
-          Text(
-            l10n.quickBattle_casualties(
-              l10n.quickBattle_attackerDefaultName,
-              result.attackerCasualties.length,
-            ),
-            style: theme.textTheme.bodySmall,
-          ),
-          Text(
-            l10n.quickBattle_casualties(
-              l10n.quickBattle_defenderDefaultName,
-              result.defenderCasualties.length,
-            ),
-            style: theme.textTheme.bodySmall,
-          ),
-          const SizedBox(height: 16),
-          Align(
-            alignment: Alignment.centerRight,
-            child: CtNinePatchButton(
-              onPressed: onDismiss,
-              child: Text(l10n.game_intervention_continue),
-            ),
-          ),
         ],
       ),
     );
