+++ tool/check_app_event_bus_decoupling.dart
@@ -81,9 +81,12 @@ const Set<String> _allowedFeatureLocalDialogFiles = <String>{
   'app/lib/features/game/flame/overlays/next_turn_confirmation_dialog.dart',
   'app/lib/features/game/screens/game/game_screen.dart',
   // Next-turn processing dialog split from `game_screen.dart` into
-  // `game_screen_fallback_next_turn.dart` for Phase 3 flame modularization.
-  // Same local-by-design rationale as the parent file (Refs #3878 Phase 3).
+  // `game_screen_fallback_next_turn.dart` and isolate runner
+  // `game_screen_fallback_next_turn_resolution.dart` for Phase 3 flame
+  // modularization. Same local-by-design rationale as the parent file
+  // (Refs #3878 Phase 3).
   'app/lib/features/game/screens/game/game_screen_fallback_next_turn.dart',
+  'app/lib/features/game/screens/game/game_screen_fallback_next_turn_resolution.dart',
   // Android back / exit-to-main-menu confirm dialog extracted from
   // game_screen.dart per `SPEC/ui/in-game-shell-narrow.md` "Android back
   // confirm". Local by design — `SPEC/program/app-ui-wiring.md` line 84.
